requirements.txt
numpy 
pandas
matplotlib
seaborn 
scikit-learn
ipykernel
pyyaml
